Cree la clase MoverAnimal que implementa el Runnable, hice que los tiburones aparecieran en el pane de InicioJuego y que ya se muevan presionando el boton de Start, modifique algunos constructores y movi los botones y textos.

+++ TyperSharkApp/src/typersharkapp/Piranha.java
@@ -11,7 +11,7 @@ public class Piranha extends AnimalesMarinos
     private Image piranha;
     private ImageView iv;
 
-    public Piranha(char caracter, Image piranha, ImageView iv, double posicionX, double posicionY, double velocidad, File listadoPalabras) {
+    public Piranha(char caracter, Image piranha, ImageView iv, double posicionX, double posicionY, int velocidad, File listadoPalabras) {
         super(posicionX, posicionY, velocidad, listadoPalabras);
         this.caracter = caracter;
         this.piranha = piranha;
